fix: address 21 PR review findings from 14 agents + external reviewers
- Extract lifecycle helpers (lifecycle.py) from app.py (915->588 lines)
- Fix dispatcher resource leak on post-start failure in auto_wire_settings
- Add BuildDispatcherFn Protocol replacing Callable[..., ...]
- Split auto_wire_phase1 into helper functions (<50 lines each)
- Make set_settings_service atomic (construct locals before assigning)
- Wrap build_dispatcher() call in try/except with specific logging
- Add deferred import comment explaining import cycle
- Remove env var name from error log, redact DB path in info log
- Document all 10 params in _build_lifecycle docstring (remove D417 noqa)
- Improve _should_auto_wire comment clarity
- Remove unused caplog fixture, tighten >= 1 to == 1 assertion
- Add 8 new tests: Phase 2 rollback, dispatcher shutdown, start failure
  invariant, SettingsService creation failure, channel deduplication,
  cost tracker/message bus error paths
- Update CLAUDE.md package structure for lifecycle.py
